can/node1: disable interrupts and message reception

+++ nodes/node_1/src/modules/gpio.c
@@ -1,5 +1,5 @@
-#include <stdio.h>
 #include "gpio.h"
+#include <stdio.h>
 
 #include <assert.h>
 
@@ -13,7 +13,7 @@ void gpio_set(GPIO *gpio, enum gpio_state state)
     }
     else
     {
-        *(gpio->port) &= ~(1 << gpio->pin);
+        *(gpio->port) &= ~(uint8_t)(1 << gpio->pin);
     }
 }
